safari-600.7-branch/LayoutTests/userscripts/user-script-image-document.html

<html>
<head>
<script>
if (window.testRunner) {
    window.testRunner.dumpAsText();
    window.testRunner.addUserScript("console.log('Start Script: ' + document.location.href)", true, true);
    window.testRunner.addUserScript("console.log('End Script: ' + document.location.href)", false, true);
}
</script>
</head>
<body>
<iframe src="resources/abe.png"></iframe>
</body>
</html>
