safari-600.7-branch/LayoutTests/userscripts/user-script-image-document-expected.txt

CONSOLE MESSAGE: line 1: End Script: user-script-image-document.html
CONSOLE MESSAGE: line 1: Start Script: abe.png
CONSOLE MESSAGE: line 1: End Script: abe.png
